Set up MySQL proxy package, improve class naming and organization
