+++ .gitattributes
@@ -8,6 +8,7 @@ phpunit.xml.dist export-ignore
 wp-setup.sh export-ignore
 /.github export-ignore
 /grammar-tools export-ignore
+/packages export-ignore
 /tests export-ignore
 /wp-includes/sqlite/class-wp-sqlite-crosscheck-db.php export-ignore
 /wordpress export-ignore
