+++ packages/wp-mysql-proxy/composer.json
@@ -1,0 +1,19 @@
+{
+    "name": "wordpress/wp-mysql-proxy",
+    "type": "library",
+	"scripts": {
+		"test": "phpunit"
+	},
+    "require-dev": {
+        "phpunit/phpunit": "^8.5",
+        "symfony/process": "^5.4"
+    },
+	"autoload": {
+		"classmap": [
+			"src/"
+		],
+		"files": [
+			"../../php-polyfills.php"
+		]
+	}
+}
